requirements.txt
Django==4.2.14
djangorestframework
black==24.4.2